remove docker pruning

+++ lib/utils.sh
@@ -347,12 +347,10 @@ spinner() {
 	local spinPidFile="${TMP_DIR}/.spinner-pid"
 	case "${action}" in
 	start)
-		test -f "${spinPidFile}" &&
-			rm "${spinPidFile}"
+		test -f "${spinPidFile}" && rm "${spinPidFile}"
 
 		# don't want to clutter logs if running headless
-		test "${HEADLESS}" == '1' &&
-			return
+		test "${HEADLESS}" == '1' && return
 
 		_start_spinner &
 		echo $! >"${spinPidFile}"
